Fixing detekt issues.
